Add find_transaction driver
This code adds a new driver to conduit-braintree to allow us to
find a transaction. It uses the Braintree::Transaction.find() method,
which takes in a parameter of the transaction id or in our case,
the reference number. Once you have the transaction you can check any
of the attributes on it such as the status.

This code also includes a includes adding the driver, adding a test
for the driver and fixtures for the test.

Remove NotFoundError